requirements.txt
dj-database-url==0.5.0
django-configurations==2.0.0
django-debug-toolbar==1.9.1
django-extensions==1.9.7
Django==2.0
django-pipeline==1.6.14
django-recaptcha==1.4.0
django-bootstrap-form==3.4
gunicorn==19.7.1
psycopg2==2.7.3.2
python-dotenv
raven==6.7.0
sendgrid==5.3.0
web3
